Abandon LaTeX in favor of AsciiDoc
While we've lost precise control over the output, there's a lot less
noise in the files and we've gained a cheap way of producing user
documentation in the HTML format.

Apparently I was eyeing DocBook all those years ago as well.  I /think/
it proved to be a little bit too hard to write that way, or to get
visually satisfying results.  Nonetheless, the raw format is capable of
specifying figures with multiple pictures, so we might want to revisit
the idea some other day.

The documents have been slightly updated to reflect the "recent" changes
in program function and hosting.  Some issues have also been corrected
in the English translation.